+++ src/libs/Share/index.ios.js
@@ -1,50 +1,11 @@
-import {AppState} from 'react-native';
-import {exists, moveFile, pathForGroup, unlink} from 'react-native-fs';
-import Onyx from 'react-native-onyx';
+import {moveFile, pathForGroup} from 'react-native-fs';
 import {ShareMenuReactView} from 'react-native-share-menu';
 import CONST from '../../CONST';
-import ONYXKEYS from '../../ONYXKEYS';
 import Navigation from '../Navigation/Navigation';
+import * as ShareActions from '../actions/Share';
 import hasNoShareData from './hasNoShareData';
 import navigateToShare from './navigateToShare';
 import normalizeShareData from './normalizeShareData';
-
-let appGroupPath;
-pathForGroup(CONST.IOS_APP_GROUP).then((path) => (appGroupPath = path));
-
-const cleanUpActions = (file) => {
-    if (!file || !file.source.includes(appGroupPath)) return [];
-    return [
-        {
-            onyxMethod: Onyx.METHOD.MERGE,
-            key: ONYXKEYS.TEMP_FILES_TO_DELETE,
-            value: [file],
-        },
-    ];
-};
-
-let isCleaningUpTempFiles = false;
-AppState.addEventListener('change', () => {
-    const connectionID = Onyx.connect({
-        key: ONYXKEYS.TEMP_FILES_TO_DELETE,
-        callback: (val) => {
-            Onyx.disconnect(connectionID);
-            if (val && !isCleaningUpTempFiles) {
-                isCleaningUpTempFiles = true;
-                val.forEach((file) => {
-                    exists(file.source).then((fileExists) => {
-                        if (!fileExists) return;
-                        unlink(file.source);
-                    });
-                });
-                // TODO: Move to actions
-                // eslint-disable-next-line rulesdir/prefer-actions-set-data
-                Onyx.set(ONYXKEYS.TEMP_FILES_TO_DELETE, []);
-                isCleaningUpTempFiles = false;
-            }
-        },
-    });
-});
 
 const dismiss = () => ShareMenuReactView.dismissExtension();
 
@@ -71,7 +32,7 @@ const registerListener = () => {
 };
 
 export default {
-    cleanUpActions,
+    cleanUpActions: ShareActions.cleanUpActions,
     dismiss,
     isShareExtension: ShareMenuReactView.isExtension,
     openApp: ShareMenuReactView.openApp,
